Moved util class and removed unused methods
