+++ common-service-core/src/main/java/com/ca/mfaas/utils/ObjectUtil.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package com.ca.mfaas.util;
+package com.ca.mfaas.utils;
 
 import lombok.experimental.UtilityClass;
