+++ common-service-core/src/test/java/com/ca/mfaas/utils/UrlUtilsTest.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package com.ca.mfaas.eurekaservice.client.util;
+package com.ca.mfaas.utils;
 
 import com.ca.mfaas.utils.UrlUtils;
 import org.junit.Test;
